Update get-instant-query-response-time.md
line 90: run-on sentence
line 93: correct word form
line 98: I think the 2nd sentence may still be a run-on, but not certain. Could just be a period after "...complete." and then new sentence starting with "This makes them..."
line 98: seems like final sentence may need to be reworked. I'm not sure what part "It's" is referring to in the sentence. 
> It’s already there, indexed by the required join key, and waits to be queried.
line 105: 2nd sentence was a run-on.