Update documentation
Yuval Boker(software engineer FB) - "with inserts you don't need to reload the join index. We know how to trigger an insert to the join index when you insert to the dimension table and the documentation should get updated."
I also have checked that, and it is true, JOIN INDEX updates by itself when you inserting data.